moved auto requirements checking to base_component
